Pull the elasticsearch client into its own file.
