+++ lib/elasticsearch/index.rb
@@ -6,6 +6,7 @@ require "rest-client"
 require "multi_json"
 require "json"
 require "elasticsearch/advanced_search_query_builder"
+require "elasticsearch/client"
 
 module Elasticsearch
   class Index
@@ -19,82 +20,6 @@ module Elasticsearch
     MASSIVE_NUMBER = 200_000
 
     attr_reader :mappings, :index_name
-
-    class Client
-
-      attr_reader :index_name  # The admin wrapper needs to get to this
-
-      # Sub-paths almost certainly shouldn't start with leading slashes,
-      # since this will make the request relative to the server root
-      SAFE_ABSOLUTE_PATHS = ["/_bulk", "/_status", "/_cluster/health"]
-
-      def initialize(base_uri, index_name, logger = nil)
-        @index_uri = base_uri + "#{CGI.escape(index_name)}/"
-        @index_name = index_name
-
-        @logger = logger || Logger.new("/dev/null")
-      end
-
-      def recording_elastic_error(&block)
-        yield
-      rescue Errno::ECONNREFUSED, Timeout::Error, SocketError
-        Rummager.statsd.increment("elasticsearcherror")
-        raise
-      end
-
-      def logging_exception_body(&block)
-        yield
-      rescue RestClient::InternalServerError => error
-        @logger.error(
-          "Internal server error in elasticsearch. " +
-          "Response: #{error.http_body}"
-        )
-        raise
-      end
-
-      def request(method, sub_path, payload)
-        recording_elastic_error do
-          logging_exception_body do
-            RestClient::Request.execute(
-              method: method,
-              url:  url_for(sub_path),
-              payload: payload,
-              headers: {content_type: "application/json"}
-            )
-          end
-        end
-      end
-
-      # Forward on HTTP request methods, intercepting and resolving URLs
-      [:get, :post, :put, :head, :delete].each do |method_name|
-        define_method method_name do |sub_path, *args|
-          full_url = url_for(sub_path)
-          @logger.debug "Sending #{method_name.upcase} request to #{full_url}"
-          args.each_with_index do |argument, index|
-            @logger.debug "Argument #{index + 1}: #{argument.inspect}"
-          end
-          recording_elastic_error do
-            logging_exception_body do
-              RestClient.send(method_name, url_for(sub_path), *args)
-            end
-          end
-        end
-      end
-
-    private
-      def url_for(sub_path)
-        if sub_path.start_with? "/"
-          path_without_query = sub_path.split("?")[0]
-          unless SAFE_ABSOLUTE_PATHS.include? path_without_query
-            @logger.error "Request sub-path '#{sub_path}' has a leading slash"
-            raise ArgumentError, "Only whitelisted absolute paths are allowed"
-          end
-        end
-
-        # Addition on URLs does relative resolution
-        (@index_uri + sub_path).to_s
-      end
-    end
 
     def initialize(base_uri, index_name, mappings, logger = nil)
       @client = Client.new(base_uri, index_name, logger)
